+++ x/delayedack/genesis.go
@@ -20,6 +20,6 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 // ExportGenesis returns the module's exported genesis
 func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
 	return &types.GenesisState{
-		RollappPackets: k.GetAllRollappPackets(ctx),
+		RollappPackets: k.ListRollappPackets(ctx, keeper.AllRollappPackets()),
 	}
 }